CHIP v0.14: 選擇權小數履約價修正 + upsert 批內去重
- taifex_options: strike 改 round(float(),2)——個股選擇權小數履約價(17.5等)
  舊 int() 全塌成 0 致同批衝突鍵重複、upsert 崩（8/12 實踩）
- models/raw: RawOptionsChain.strike Integer→Float（SQLite 鬆型別免遷移）
- base.upsert: .values() 前按 conflict_cols 批內去重（通用安全網）
- 另含 taifex_options._fetch_json utf-8-sig（吃 BOM，TAIFEX 8/10-11 暫時性 CSV 事件）
- 8/12 chain 13,059 筆補齊；歷史 strike=0 殘留 2,240 筆不可回補（端點僅供最新）

+++ backend/app/models/raw.py
@@ -9,13 +9,14 @@ class RawOptionsChain(Base):
     __tablename__ = "raw_options_chain"
     __table_args__ = (
         UniqueConstraint("date", "contract", "expiry", "strike", "call_put", "trading_session", name="uq_opt_chain"),
+        # strike 為 Float：個股選擇權有小數履約價（17.5 等），SQLite 欄位鬆型別、舊整數值不需遷移（2026-08-13）
     )
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     date = Column(Date, nullable=False, index=True)
     contract = Column(String(20), nullable=False, index=True)
     expiry = Column(String(20), nullable=False)
-    strike = Column(Integer, nullable=False)
+    strike = Column(Float, nullable=False)
     call_put = Column(String(4), nullable=False)
     trading_session = Column(String(10), nullable=False, default="一般")  # 一般（日盤）/ 盤後（夜盤）
     open = Column(Float)
